增加成就达成

+++ Core/MsgHandler/CountryInnerMsgHandler.h
@@ -704,7 +704,7 @@ void OnClanSkillDonateCheckReq( GameMsgHdr& hdr, const void * data )
 	struct DonateItems
 	{
 		UInt8  skillId;
-		UInt16 flag;
+		UInt8 flag;
 		UInt32 count;
 		UInt8 ret;
 	};
