+++ Core/MsgHandler/WorldOuterMsgHandler.h
@@ -654,7 +654,7 @@ void OnClanTechOpReq(GameMsgHdr& hdr, const void * data)
 				struct DonateItems
 				{
 					UInt8  skillId;
-					UInt16 flag;
+					UInt8 flag;
 					UInt32 count;
 					UInt8 ret;
 				};
